@@ -3,47 +3,46 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Franca
+namespace Franca;
+
+public sealed class ChoiceTokenizer : ITokenizer
 {
-	public sealed class ChoiceTokenizer : ITokenizer
+	public ChoiceTokenizer(IEnumerable<ITokenizer> choices)
 	{
-		public ChoiceTokenizer(IEnumerable<ITokenizer> choices)
-		{
-			this.Choices = choices;
-		}
+		this.Choices = choices;
+	}
 
-		public IEnumerable<ITokenizer> Choices { get; }
+	public IEnumerable<ITokenizer> Choices { get; }
 
-		public Token Parse(ReadOnlySpan<char> span)
+	public Token Parse(ReadOnlySpan<char> span)
+	{
+		var ambiguous = new List<ITokenizer>();
+		foreach (var parser in this.Choices)
 		{
-			var ambiguous = new List<ITokenizer>();
-			foreach (var parser in this.Choices)
+			var result = parser.Parse(span);
+			if (result.IsSuccess)
 			{
-				var result = parser.Parse(span);
-				if (result.IsSuccess)
-				{
-					return result;
-				}
+				return result;
 			}
-
-			return Token.Fail(span);
 		}
 
-		public override string ToString() => "( " + string.Join(" | ", this.Choices.Select(i => i.ToString())) + " )";
+		return Token.Fail(span);
+	}
 
-		public static ChoiceTokenizer operator |(ChoiceTokenizer left, ChoiceTokenizer right)
-		{
-			return new ChoiceTokenizer(left.Choices.Concat(right.Choices));
-		}
+	public override string ToString() => "( " + string.Join(" | ", this.Choices.Select(i => i.ToString())) + " )";
 
-		public static ChoiceTokenizer operator |(ChoiceTokenizer left, ITokenizer right)
-		{
-			return new ChoiceTokenizer(left.Choices.Append(right));
-		}
+	public static ChoiceTokenizer operator |(ChoiceTokenizer left, ChoiceTokenizer right)
+	{
+		return new ChoiceTokenizer(left.Choices.Concat(right.Choices));
+	}
 
-		public static ChoiceTokenizer operator |(ITokenizer left, ChoiceTokenizer right)
-		{
-			return new ChoiceTokenizer(new[] { left }.Concat(right.Choices));
-		}
+	public static ChoiceTokenizer operator |(ChoiceTokenizer left, ITokenizer right)
+	{
+		return new ChoiceTokenizer(left.Choices.Append(right));
+	}
+
+	public static ChoiceTokenizer operator |(ITokenizer left, ChoiceTokenizer right)
+	{
+		return new ChoiceTokenizer(new[] { left }.Concat(right.Choices));
 	}
 }
